Reorganize mem functions
This reduces the amount of platform-specific code
